requirements.txt
Django==2.0.5
psycopg2
python-ldap
uwsgi
requests